Turned on a couple of additional gcc warnings, and fixed some minor issues found by those warnings.

+++ iTetPreferencesWindowController.m
@@ -55,7 +55,7 @@
 - (void)displayViewControllerAtIndex:(iTetPreferencesTabNumber)index
 {
 	// Sanity check
-	NSParameterAssert((index >= 0) && (index < [viewControllers count]));
+	NSParameterAssert((index >= 0) && (((NSUInteger)index) < [viewControllers count]));
 	
 	// Check if we are already displaying the view
 	if (currentViewNumber == index)
